Fix several special warnings.

+++ msg.c
@@ -1,7 +1,5 @@
 /**
- * @file msg.c Message functions
- *
- * purple
+ * Copyright (C) 2007 Felipe Contreras
  *
  * Purple is the legal property of its developers, whose names are too numerous
  * to list here.  Please refer to the COPYRIGHT file distributed with this
@@ -758,7 +756,7 @@ msn_message_show_readable(MsnMessage *msg, const char *info,
 			}
 			else
 			{
-				int i;
+				gsize i;
 				for (i = 0; i < msg->body_len; i++)
 				{
 					g_string_append_printf(str, "%.2hhX ", body[i]);
